requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.1
Flask-Migrate==1.6.0
blinker==1.4
Flask-WTF==0.12
WTForms==2.1
python-slugify==1.1.4
